Replace continuous=yes + sidewalk=yes with highway=crossing
TfL's side road entry treatments always represent a crossing and importing them without highway=crossing provides no benefit to data consumers.

Only a minority of these crossings are "Copenhagen crossings" (the term used by London Borough of Waltham Forest), which have a continuous and unmarked sidewalk across the side road and usually no tactile paving. They also have a give way marking on the side road before it crosses the sidewalk, giving priority to pedestrians crossing the side road. In these cases, the continuous=yes tag might be useful when documented in the wiki with an accepted definition.
https://enjoywalthamforest.co.uk/blended-copenhagen-crossings/

The majority are actually normal pedestrian crossings with a speed table or flush kerbs. The pavement is not continuous, the give way markings on the side road are between the crossing and the main road. Tactile paving is often present. Telling a data consumer with additional accessibility needs that such a side road treatment is a continuous sidewalk would have been somewhat irresponsible.